worker/requirements.txt
fastapi[all]==0.112.1
redis==5.2.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
openai-whisper==20240930